+++ tests/manual/test-wrappers.R
@@ -2,7 +2,7 @@
 # if you test the package using `pkgload::load_all`
 
 d <- tibble::tibble(
-  x = seq(0, 1e6),
+  x = seq(0, 1e8),
   t = nanotime::nanotime(Sys.time()) + seq(0, 1e6) * 7e4,
   tp = Sys.time() + seq(0, 1e6) * 7,
   tch = format(t, "%Y-%m-%d %H:%M:%E9S"),
@@ -11,26 +11,26 @@ d <- tibble::tibble(
 
 # data downsampler
 
-apply_downsampler(d$y)
-apply_downsampler(d)
-apply_downsampler(d, n_out = 100, aggregator = min_max_aggregator)
-apply_downsampler(d, n_out = 100, aggregator = range_stat_aggregator)
+shiny_hugeplot(d$y)
+shiny_hugeplot(d)
+shiny_hugeplot(d, aggregator = range_stat_aggregator$new(), n_out = 100)
+shiny_hugeplot(d, aggregator = range_stat_aggregator$new(y = NULL), n_out = 100)
 
 # plotly downsampler
 
 plotly::plot_ly() %>%
   add_trace(x = d$x, y = d$y, type = "scatter", mode = "lines") %>%
-  apply_downsampler()
+  shiny_hugeplot()
 
 
 plotly::plot_ly() %>%
   add_trace(x = d$x, y = d$y, type = "scatter", mode = "lines") %>%
-  apply_downsampler(n_out = 100, aggregator = range_stat_aggregator)
+  shiny_hugeplot(n_out = 100, aggregator = range_stat_aggregator)
 
 plotly::plot_ly() %>%
   add_trace(x = d$t, y = d$y, type = "scatter", mode = "lines") %>%
   layout(xaxis = list(type = "date")) %>%
-  apply_downsampler()
+  shiny_hugeplot()
 
 # subplot downsampler
 
@@ -56,13 +56,13 @@ p4 <- plotly::plot_ly(
 
 ps <- subplot(p1, p2, p3, p4, nrows = 2)
 
-apply_downsampler(ps)
+shiny_hugeplot(ps, n_out = 100)
 
 
 # subplot downsampler (x axis is shared)
 
 ps <- subplot(p1, p2, p3, p4, nrows = 2, shareX = TRUE) %>%
-  apply_downsampler()
+  shiny_hugeplot()
 
 # subplot downsampler (x axis is shared using `matches`)
 
@@ -70,28 +70,6 @@ ps <- subplot(p1, p2, p3, p4, nrows = 2) %>%
   layout(
     xaxis = list(matches = "x4")
   ) %>%
-  apply_downsampler()
+  shiny_hugeplot()
 
 
-
-
-# dash downsampler
-# app <- dash::dash_app() %>%
-#   dash::set_layout(
-#     dash::dccGraph(
-#       id = "downsample-figure",
-#       figure = plotly::plot_ly(
-#         data = d[1:1e4, ], x = ~x, y = ~y, type = "scatter", mode = "lines"
-#       )
-#     ),
-#     dash::dccGraph(
-#       id = "other-figure",
-#       figure = plotly::plot_ly(
-#         data = d[1e2 + 1:1e2, ], x = ~x, y = ~y,
-#         type = "scatter", mode = "lines"
-#       )
-#     )
-#   ) %>%
-#   apply_downsampler(graph_id = "downsample-figure")
-#
-# app$run_server(host = "127.0.0.1", port = "8050", use_viewer = TRUE)