+++ NEWS.md
@@ -1,3 +1,8 @@
+# shinyHugePlot 0.1.0
+
+* Interactive change of the aggregation algorithm is introduced
+* Names of the multiple series can be given with formula
+
 # shinyHugePlot 0.0.1
 
 * The first version.